Working complete CRUD for state

+++ project/project/src/main/java/com/epam/spring/framework/variant/timekeeping/repository/StateRepository.java
@@ -3,5 +3,5 @@ package com.epam.spring.framework.variant.timekeeping.repository;
 import com.epam.spring.framework.variant.timekeeping.domain.State;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StateRepository extends CrudRepository<State,Long> {
+public interface StateRepository extends CrudRepository<State,Integer> {
 }
